backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
aiosqlite==0.20.0
python-dotenv==1.0.1
google-auth==2.37.0
httpx==0.28.1
aiosmtplib==3.0.2
